requirements.txt
blinker==1.8.0
click==8.1.7
exceptiongroup==1.2.1
Flask==3.0.3
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
pluggy==1.5.0
pytest==8.2.0
python-dotenv==1.0.1
tomli==2.0.1
Werkzeug==3.0.2
zipp==3.18.1
